requirements.txt
pandas==2.1.0
nltk==3.8.1
pymongo==4.2.0
python-dotenv==1.0.0
twikit==0.2.0
